[hotfix] [tests] Move utility methods into correct test class.
The methods were intended to be generic for tests across partitions and hence placed in the
SubpartitionTestBase.

However, the SubpartitionTestBase can only really test common contract behavior, like behavior on
disposal, finishing, and buffer reycling contracts in those cases. Producer / consumer behavior
is sufficiently different between both implementations that it does not make sense at this point
to try and share the tests.
