+++ flink-runtime/src/test/java/org/apache/flink/runtime/io/network/partition/SubpartitionTestBase.java
@@ -18,27 +18,18 @@
 
 package org.apache.flink.runtime.io.network.partition;
 
-import org.apache.flink.runtime.event.AbstractEvent;
-import org.apache.flink.runtime.io.network.api.serialization.EventSerializer;
 import org.apache.flink.runtime.io.network.buffer.BufferBuilderTestUtils;
 import org.apache.flink.runtime.io.network.buffer.BufferConsumer;
 import org.apache.flink.util.TestLogger;
 
 import org.junit.Test;
 
-import javax.annotation.Nullable;
-
-import java.io.IOException;
-
 import static org.apache.flink.runtime.io.network.buffer.BufferBuilderTestUtils.createFilledBufferConsumer;
-import static org.apache.flink.util.Preconditions.checkArgument;
-import static org.hamcrest.core.IsInstanceOf.instanceOf;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 
 /**
@@ -134,85 +125,5 @@ public abstract class SubpartitionTestBase extends TestLogger {
 
 		// Verify that parent release is reflected at partition view
 		assertTrue(view.isReleased());
-	}
-
-	static void assertNextBuffer(
-			ResultSubpartitionView readView,
-			int expectedReadableBufferSize,
-			boolean expectedIsMoreAvailable,
-			int expectedBuffersInBacklog,
-			boolean expectedNextBufferIsEvent,
-			boolean expectedRecycledAfterRecycle) throws IOException, InterruptedException {
-		assertNextBufferOrEvent(
-			readView,
-			expectedReadableBufferSize,
-			true,
-			null,
-			expectedIsMoreAvailable,
-			expectedBuffersInBacklog,
-			expectedNextBufferIsEvent,
-			expectedRecycledAfterRecycle);
-	}
-
-	static void assertNextEvent(
-			ResultSubpartitionView readView,
-			int expectedReadableBufferSize,
-			Class<? extends AbstractEvent> expectedEventClass,
-			boolean expectedIsMoreAvailable,
-			int expectedBuffersInBacklog,
-			boolean expectedNextBufferIsEvent,
-			boolean expectedRecycledAfterRecycle) throws IOException, InterruptedException {
-		assertNextBufferOrEvent(
-			readView,
-			expectedReadableBufferSize,
-			false,
-			expectedEventClass,
-			expectedIsMoreAvailable,
-			expectedBuffersInBacklog,
-			expectedNextBufferIsEvent,
-			expectedRecycledAfterRecycle);
-	}
-
-	private static void assertNextBufferOrEvent(
-			ResultSubpartitionView readView,
-			int expectedReadableBufferSize,
-			boolean expectedIsBuffer,
-			@Nullable Class<? extends AbstractEvent> expectedEventClass,
-			boolean expectedIsMoreAvailable,
-			int expectedBuffersInBacklog,
-			boolean expectedNextBufferIsEvent,
-			boolean expectedRecycledAfterRecycle) throws IOException, InterruptedException {
-		checkArgument(expectedEventClass == null || !expectedIsBuffer);
-
-		ResultSubpartition.BufferAndBacklog bufferAndBacklog = readView.getNextBuffer();
-		assertNotNull(bufferAndBacklog);
-		try {
-			assertEquals("buffer size", expectedReadableBufferSize,
-				bufferAndBacklog.buffer().readableBytes());
-			assertEquals("buffer or event", expectedIsBuffer,
-				bufferAndBacklog.buffer().isBuffer());
-			if (expectedEventClass != null) {
-				assertThat(EventSerializer
-						.fromBuffer(bufferAndBacklog.buffer(), ClassLoader.getSystemClassLoader()),
-					instanceOf(expectedEventClass));
-			}
-			assertEquals("more available", expectedIsMoreAvailable,
-				bufferAndBacklog.isMoreAvailable());
-			assertEquals("more available", expectedIsMoreAvailable, readView.isAvailable());
-			assertEquals("backlog", expectedBuffersInBacklog, bufferAndBacklog.buffersInBacklog());
-			assertEquals("next is event", expectedNextBufferIsEvent,
-				bufferAndBacklog.nextBufferIsEvent());
-			assertEquals("next is event", expectedNextBufferIsEvent,
-				readView.nextBufferIsEvent());
-
-			assertFalse("not recycled", bufferAndBacklog.buffer().isRecycled());
-		} finally {
-			bufferAndBacklog.buffer().recycleBuffer();
-		}
-		assertEquals("recycled", expectedRecycledAfterRecycle, bufferAndBacklog.buffer().isRecycled());
-	}
-
-	static void assertNoNextBuffer(ResultSubpartitionView readView) throws IOException, InterruptedException {
-		assertNull(readView.getNextBuffer());
 	}
 }
